requirements.txt
fastapi
uvicorn
jinja2
weasyprint
sqlalchemy
python-dotenv
httpx
fpdf2
